SS-8135: separated session and viewport store
