+++ hooks/shapediver/viewer/interaction/useConvertDraggingData.ts
@@ -1,5 +1,5 @@
 import { convertUserDefinedNameFilters, OutputNodeNameFilterPatterns } from "@shapediver/viewer.features.interaction";
-import { useShapeDiverStoreViewer } from "../../../../store/useShapeDiverStoreViewer";
+import { useShapeDiverStoreSession } from "../../../../store/useShapeDiverStoreSession";
 import { useState, useEffect } from "react";
 import { IDraggingParameterProps } from "@shapediver/viewer";
 import { vec3 } from "gl-matrix";
@@ -40,7 +40,7 @@ export function useConvertDraggingData(sessionIds?: string[], draggingProps?: ID
 	objects: ConvertedDragObject[]
 } {
 	// get the session API
-	const sessions = useShapeDiverStoreViewer(state => { return state.sessions; });
+	const sessions = useShapeDiverStoreSession(state => { return state.sessions; });
 
 	// create a state for the pattern
 	const [objects, setObjects] = useState<ConvertedDragObject[]>([]);
