+++ hooks/shapediver/viewer/useViewport.ts
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
-import { useShapeDiverStoreViewer } from "../../../store/useShapeDiverStoreViewer";
-import { ViewportCreateDto } from "../../../types/store/shapediverStoreViewer";
+import { useShapeDiverStoreViewport } from "../../../store/useShapeDiverStoreViewport";
+import { ViewportCreateDto } from "../../../types/store/shapediverStoreViewport";
 import { useShallow } from "zustand/react/shallow";
 import { useViewportId } from "./useViewportId";
 
@@ -8,12 +8,12 @@ import { useViewportId } from "./useViewportId";
  * Hook for creating a viewport of the ShapeDiver 3D Viewer.
  * Typically, you want to directly use the {@link ViewportComponent} instead
  * of calling this hook yourself.
- * @see {@link useShapeDiverStoreViewer} to access the API of the viewport.
+ * @see {@link useShapeDiverStoreViewport} to access the API of the viewport.
  * @param props
  * @returns
  */
 export function useViewport(props: ViewportCreateDto) {
-	const { createViewport, closeViewport } = useShapeDiverStoreViewer(
+	const { createViewport, closeViewport } = useShapeDiverStoreViewport(
 		useShallow(state => ({ createViewport: state.createViewport, closeViewport: state.closeViewport }))
 	);
 	const [error, setError] = useState<Error | undefined>(undefined);
